agent_interface/requirements.txt
flask==2.0.1
werkzeug==2.0.1
requests==2.26.0
python-dotenv==0.19.0
google-generativeai==0.3.0
selenium==4.1.0
beautifulsoup4==4.9.3
pytest==6.2.5
pytest-html==3.1.1 